Ignore http content-types when finding feed urls
Because of badly behaving web servers http content-type headers might
not be accurate for atom/rss feeds. So we don't trust content-type
headers anymore. Now we try to parse everything as proper feed and if
that fails we can be actually be sure we have something else in our
hands.